biome projectiles improvements

+++ Gambler/Projectiles/DesertCardProj.cs
@@ -25,6 +25,7 @@ namespace OrchidMod.Gambler.Projectiles
 			Projectile.aiStyle = 0;
 			Projectile.timeLeft = 30;
 			Projectile.penetrate = 2;
+			Projectile.alpha = 255;
 			this.gamblingChipChance = 10;
 		}
 		
@@ -70,7 +71,7 @@ namespace OrchidMod.Gambler.Projectiles
 		public void checkMouseDrag() {
 			Projectile proj = Main.projectile[(int)Projectile.ai[0]];
 			
-			if (proj.type != ProjectileType<Gambler.Projectiles.DesertCardBase>() || proj.active == false) {
+			if (proj.type != ProjectileType<Gambler.Projectiles.DesertCardBase>() || proj.active == false && Projectile.ai[1] != 2f) {
 				Projectile.Kill();
 			}
 			
@@ -118,6 +119,7 @@ namespace OrchidMod.Gambler.Projectiles
 						Projectile.tileCollide = true;
 						Projectile.friendly = true;
 						Projectile.netUpdate = true;
+						Projectile.alpha = 0;
 						SoundEngine.PlaySound(SoundID.Item1, Projectile.Center);
 					}
 				}
